+++ apps/server/src/constants.ts
@@ -1,4 +1,4 @@
-export const PUBLIC_ROUTES = ["/api/v1/auth", "/api/v1/stats"];
+export const PUBLIC_ROUTES = ["/api/v1/auth", "/api/v1/stats", "/api/v1/docs", "/api/v1/openapi.yaml"];
 export const APP_NAME = process.env.APP_NAME || "OpenBTS";
 export const API_KEYS_LIMIT = 1;
 export const API_KEY_COOLDOWN_SECONDS = 60 * 60 * 24 * 7; // 7 days